+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
  - v2.3 Enumerable#chunk_while and v2.2 Enumerable#slice_after should not require `count`
  - v2.4 Array#sum should not use `each` (the Enumerable version does)
+ - load modules before classes so they don't override the local method
 
 # [0.7.0][] (2017-04-22)